move command struct to own module
