+++ gerritbot/src/lib.rs
@@ -3,7 +3,6 @@ use std::convert;
 use std::fs::File;
 use std::io;
 use std::path::Path;
-use std::str::FromStr;
 use std::time::Duration;
 
 use futures::{future::Future, stream, stream::Stream};
@@ -15,10 +14,12 @@ use gerritbot_gerrit as gerrit;
 use gerritbot_spark as spark;
 
 pub mod args;
+mod command;
 mod format;
 mod rate_limit;
 mod state;
 
+use command::Command;
 use format::Formatter;
 pub use format::DEFAULT_FORMAT_SCRIPT;
 use rate_limit::RateLimiter;
@@ -102,44 +103,6 @@ impl Builder {
             formatter,
             state,
         }
-    }
-}
-
-#[derive(Debug)]
-enum Command {
-    Enable,
-    Disable,
-    Status,
-    Help,
-    Version,
-    FilterStatus,
-    FilterEnable(bool),
-    FilterAdd(String),
-}
-
-impl FromStr for Command {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        lazy_static! {
-            static ref FILTER_REGEX: Regex = Regex::new(r"(?i)^filter (.*)$").unwrap();
-        };
-
-        Ok(match &s.trim().to_lowercase()[..] {
-            "enable" => Command::Enable,
-            "disable" => Command::Disable,
-            "status" => Command::Status,
-            "help" => Command::Help,
-            "version" => Command::Version,
-            "filter" => Command::FilterStatus,
-            "filter enable" => Command::FilterEnable(true),
-            "filter disable" => Command::FilterEnable(false),
-            _ => FILTER_REGEX
-                .captures(&s.trim()[..])
-                .and_then(|cap| cap.get(1))
-                .map(|m| Command::FilterAdd(m.as_str().to_string()))
-                .ok_or(())?,
-        })
     }
 }
